some refactoring adds local storage.js
