+++ lib/IdeaBox.js
@@ -1,0 +1,9 @@
+function IdeaBox(title, idea, id){
+  this.title = title;
+  this.idea = idea;
+  this.id = id;
+  this.quality = "swill";
+};
+
+
+module.exports = IdeaBox;
