+++ lib/buttons.js
@@ -1,4 +1,7 @@
 // const qualitySheet = require('./qualities.js')
+const IdeaBox = require('./IdeaBox.js')
+const createIdeaBox = require('./createToDo.js')
+const storage = require('./localstorage.js')
 
 
 const buttons = {
@@ -6,11 +9,10 @@ const buttons = {
    $(".save-btn").on("click", () => {
      let title = $(".title-input").val();
      let idea = $(".idea-input").val();
-     let timeStamp = getTimeStamp();
-     let ideabox = new IdeaBox(title, idea, Date.now(), timeStamp);
+     let ideabox = new IdeaBox(title, idea, Date.now());
      let key = ideabox.id;
-     localStorage.setItem(key, JSON.stringify(ideabox));
-     createIdeaBox(ideabox, timeStamp);
+     storage.storeIdea(key, ideabox);
+     createIdeaBox(ideabox);
      emptyInput();
      $(".title-input").focus();
    })
@@ -25,7 +27,7 @@ const buttons = {
      let ideabox = JSON.parse(localStorage.getItem(key));
      let newQuality = getNewQuality(selector, quality.text());
      ideabox.quality = newQuality;
-     localStorage.setItem(key, JSON.stringify(ideabox));
+     storage.storeIdea(key, ideabox);
      quality.text(newQuality);
    })
  },
@@ -82,6 +84,12 @@ function downVote(quality){
     default:
       return "swill";
   }
+}
+
+function emptyInput() {
+  $(".title-input").val("");
+  $(".idea-input").val("");
+  $(".idea-input").css("height", "42px");
 }
 
 module.exports = buttons;
