+++ lib/index.js
@@ -4,12 +4,15 @@ const sass = require('./styles.scss');
 const css = require('../css/styles.css')
 const createIdeaBox = require('./createToDo.js')
 const buttons = require('./buttons.js')
+const storage = require('./localstorage.js')
+const IdeaBox = require('./IdeaBox.js')
 
 
 buttons.saveButton();
 buttons.voteQuality();
 buttons.deleteButton();
 buttons.searchValue();
+
 
 let sortCount = 0;
 
@@ -19,14 +22,6 @@ $(() => {
     createIdeaBox(storedIdeaBox);
   }
 });
-
-function IdeaBox(title, idea, id, timeStamp){
-  this.title = title;
-  this.idea = idea;
-  this.id = id;
-  this.quality = "swill";
-  this.timeStamp = timeStamp;
-};
 
 $(".idea-container").on("focus", ".idea-title, .idea-body", () => {
   let selector = $(this).closest(".idea-card");
@@ -42,9 +37,10 @@ $(".idea-container").on("focus", ".idea-title, .idea-body", () => {
   $(this).on("blur", () => {
     ideabox.title = selector.find(".idea-title").text();
     ideabox.idea = selector.find(".idea-body").text();
-    localStorage.setItem(key, JSON.stringify(ideabox));
+    storage.storeIdea(key, ideabox);
   })
 })
+
 
 $(".title-input, .idea-input").on("keyup", () => {
   /\S/.test($(".title-input").val()) && /\S/.test($(".idea-input").val()) ? $(".save-btn").prop("disabled", false) : $(".save-btn").prop("disabled", true);
@@ -69,13 +65,7 @@ $(".input-search").on("keyup", () => {
 });
 
 
-function emptyInput() {
-  $(".title-input").val("");
-  $(".idea-input").val("");
-  $(".idea-input").css("height", "42px");
-}
-
-function getTimeStamp(){
-  let time = Date();
-    return time;
-}
+// function getTimeStamp(){
+//   let time = Date();
+//     return time;
+// }
