+++ lib/localstorage.js
@@ -1,0 +1,9 @@
+
+const storage = {
+   storeIdea: function (key, ideabox){
+    localStorage.setItem(key, JSON.stringify(ideabox));
+  },
+
+}
+
+  module.exports = storage;
